assignment_materials/05.FastAPI_agantStreaming/pyproject.toml
[project]
name = "05-fastapi-agantstreaming"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "fastapi>=0.135.3",
    "greenlet>=3.3.2",
    "openai-agents[litellm]>=0.13.4",
    "python-dotenv>=1.2.2",
    "scalar-fastapi>=1.8.1",
    "simpleeval>=1.0.7",
    "sqlmodel>=0.0.38",
    "tavily-python>=0.7.23",
    "uvicorn>=0.43.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E" ,"F","W","I"]
exclude = ["alembic"]
